+++ src/main/java/com/hcl/appscan/jenkins/plugin/scanners/ScannerConstants.java
@@ -9,7 +9,12 @@ public interface ScannerConstants {
 	String EXTRA_FIELD				= "ExtraField";				//$NON-NLS-1$
 	String INCLUDE_VERIFIED_DOMAINS			= "IncludeVerifiedDomains";		//$NON-NLS-1$
 	String LOGIN_USER				= "LoginUser";				//$NON-NLS-1$
-	String LOGIN_PASSWORD				= "LoginPassword";			//$NON-NLS-1$
+	String LOGIN_PASSWORD				= "LoginPassword";				//$NON-NLS-1$
+	String LOGIN_TYPE				= "LoginType";				//$NON-NLS-1$
+	String AUTOMATIC				= "Automatic";				//$NON-NLS-1$
+	String RECORDED					= "Manual";				//$NON-NLS-1$
+	String NONE					= "None";				//$NON-NLS-1$
+	String TRAFFIC_FILE				= "trafficFile";				//$NON-NLS-1$
 	String PRESENCE_ID				= "PresenceId";				//$NON-NLS-1$
 	String SCAN_FILE				= "ScanFile";				//$NON-NLS-1$
 	String SCAN_TYPE				= "ScanType";				//$NON-NLS-1$
@@ -17,10 +22,11 @@ public interface ScannerConstants {
 	String TARGET					= CoreConstants.TARGET;			//$NON-NLS-1$
 	String TEMPLATE_EXTENSION			= ".scant";				//$NON-NLS-1$
 	String TEMPLATE_EXTENSION2			= ".scan";				//$NON-NLS-1$
+	String TEMPLATE_EXTENSION3			= ".config";				//$NON-NLS-1$
 	
 	String DYNAMIC_ANALYZER				= "Dynamic Analyzer";			//$NON-NLS-1$
-	String MOBILE_ANALYZER				= "Mobile Analyzer";			//$NON-NLS-1$
 	String STATIC_ANALYZER				= "Static Analyzer";			//$NON-NLS-1$
+  	String SOFTWARE_COMPOSITION_ANALYZER		= "Software Composition Analyzer";	//$NON-NLS-1$
 
 	String CUSTOM					= "Custom";				//$NON-NLS-1$
 	String PRODUCTION				= "Production";				//$NON-NLS-1$
@@ -33,4 +39,5 @@ public interface ScannerConstants {
 	String FASTEST                  = "Fastest";                //$NON-NLS-1$
 	String NO_OPTIMIZATION          = "NoOptimization";         //$NON-NLS-1$
 	String TEST_OPTIMIZATION_LEVEL  = "TestOptimizationLevel";  //$NON-NLS-1$
+    	String SCAN_SPEED               = "scanSpeed";              //$NON-NLS-1$
 }